뜨고 있던 모든 경고 메시지 수정
경고 메시지가 뜨지 않게 수정

+++ Server/Network/Sender.cpp
@@ -91,6 +91,6 @@ void Sender::_SendRUI(SOCKET sock)
 
 void Sender::_Send(SOCKET sock, const char* buf)
 {
-	int strLen = strlen(buf);
+	size_t strLen = strlen(buf);
 	send(sock, buf, strLen, 0);
 }
